replace sha256pow with poscan-pow consensus crate
