+++ Cargo.toml
@@ -9,8 +9,8 @@ edition = "2024"
 members = [
     "consensus/obj-asteroid",
     "consensus/poscan",
+    "consensus/poscan-pow",
     "consensus/sc-consensus-pow",
-    "consensus/sha256pow",
     "node",
     "pallets/difficulty",
     "pallets/reward",
@@ -22,8 +22,8 @@ resolver = "2"
 
 [workspace.dependencies]
 solochain-template-runtime = { path = "./runtime", default-features = false }
-sha256pow = { path = "./consensus/sha256pow", default-features = false }
 poscan = { path = "./consensus/poscan", default-features = false }
+poscan-pow = { path = "./consensus/poscan-pow", default-features = false }
 obj-asteroid = { path = "./consensus/obj-asteroid", default-features = false }
 pallet-difficulty = { path = "./pallets/difficulty", default-features = false }
 pallet-reward = { path = "./pallets/reward", default-features = false }
