+++ node/Cargo.toml
@@ -95,7 +95,7 @@ sp-timestamp.default-features = true
 sp-timestamp.workspace = true
 substrate-frame-rpc-system.default-features = true
 substrate-frame-rpc-system.workspace = true
-sha256pow.workspace = true
+poscan-pow.workspace = true
 poscan.workspace = true
 
 # Frontier (Ethereum compatibility) — node-side RPC stack.
