+++ runtime/src/apis.rs
@@ -275,7 +275,7 @@ impl_runtime_apis! {
 		}
 	}
 
-	impl sha256pow::PowVerifyApi<Block> for Runtime {
+	impl poscan_pow::PowVerifyApi<Block> for Runtime {
 		fn verify_seal(pre_hash: sp_core::H256, seal: Vec<u8>, difficulty: U256) -> bool {
 			poscan::verify_seal(pre_hash, &seal, difficulty).unwrap_or(false)
 		}
